+++ src/tools/registry.rs
@@ -66,6 +66,12 @@ const PROTECTED_TOOL_NAMES: &[&str] = &[
     "create_relationship",
     "search_entities",
     "get_entity_context",
+    "delete_entity",
+    "merge_entities",
+    "stage_candidate",
+    "list_candidates",
+    "approve_candidate",
+    "reject_candidate",
 ];
 
 /// Registry of available tools.
@@ -250,16 +256,23 @@ impl ToolRegistry {
     #[cfg(feature = "neo4j")]
     pub fn register_graph_tools(&self, client: Arc<crate::graph::GraphClient>) {
         use crate::tools::builtin::{
-            CreateEntityTool, CreateRelationshipTool, GetEntityContextTool, SearchEntitiesTool,
-            UpdateEntityTool,
+            ApproveCandidateTool, CreateEntityTool, CreateRelationshipTool, DeleteEntityTool,
+            GetEntityContextTool, ListCandidatesTool, MergeEntitiesTool, RejectCandidateTool,
+            SearchEntitiesTool, StageCandidateTool, UpdateEntityTool,
         };
         self.register_sync(Arc::new(CreateEntityTool::new(Arc::clone(&client))));
         self.register_sync(Arc::new(UpdateEntityTool::new(Arc::clone(&client))));
         self.register_sync(Arc::new(CreateRelationshipTool::new(Arc::clone(&client))));
         self.register_sync(Arc::new(SearchEntitiesTool::new(Arc::clone(&client))));
-        self.register_sync(Arc::new(GetEntityContextTool::new(client)));
+        self.register_sync(Arc::new(GetEntityContextTool::new(Arc::clone(&client))));
+        self.register_sync(Arc::new(DeleteEntityTool::new(Arc::clone(&client))));
+        self.register_sync(Arc::new(MergeEntitiesTool::new(Arc::clone(&client))));
+        self.register_sync(Arc::new(StageCandidateTool::new(Arc::clone(&client))));
+        self.register_sync(Arc::new(ListCandidatesTool::new(Arc::clone(&client))));
+        self.register_sync(Arc::new(ApproveCandidateTool::new(Arc::clone(&client))));
+        self.register_sync(Arc::new(RejectCandidateTool::new(client)));
 
-        tracing::info!("Registered 5 knowledge graph tools");
+        tracing::info!("Registered 11 knowledge graph tools");
     }
 
     /// Register the web search tool with the given backend.